backend/pyproject.toml
[project]
name = "padel-video-coach-backend"
version = "0.1.0"
description = "FastAPI backend for padel video analysis jobs"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "pydantic-settings>=2.7.0",
  "python-multipart>=0.0.12",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0",
  "ruff>=0.8.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
